Dashboard/charging-vehicle/vehicleSpecs implemented

+++ src/main/java/com/stephanie/ev_charging_pro/model/Vehicle.java
@@ -20,7 +20,7 @@ public class Vehicle {
     private String model;
     private String plateNumber;
 
-    private int batteryCapacity;
+    private double batteryCapacity;
     private int currentBatteryLevel;
 
     @ManyToOne
